requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
requests
websockets
stomp.py
